changed cli, added Kraken3Base class. merged back drivers into kraken3
