+++ liquidctl/driver/commander_pro.py
@@ -564,6 +564,6 @@ class CommanderPro(UsbHidDriver):
         buf = bytes(self.device.read(_RESPONSE_LENGTH))
         return buf
 
-    def set_screen(self, mode, value, **kwargs):
+    def set_screen(self, channel, mode, value, **kwargs):
         """Not supported by this device."""
         raise NotSupportedByDevice()
