+++ liquidctl/driver/kraken2.py
@@ -291,7 +291,7 @@ class Kraken2(UsbHidDriver):
         padding = [0x0]*(_WRITE_LENGTH - len(data))
         self.device.write(data + padding)
 
-    def set_screen(self, mode, value, **kwargs):
+    def set_screen(self, channel, mode, value, **kwargs):
         """Not supported by this device."""
         raise NotSupportedByDevice()
