+++ liquidctl/driver/nzxt_epsu.py
@@ -145,7 +145,7 @@ class NzxtEPsu(UsbHidDriver):
         ascam_ver = int.from_bytes(bytes.fromhex(human_ver), byteorder='big')
         return (human_ver, ascam_ver)
 
-    def set_screen(self, mode, value, **kwargs):
+    def set_screen(self, channel, mode, value, **kwargs):
         """Not supported by this device."""
         raise NotSupportedByDevice()
