+++ liquidctl/driver/smart_device.py
@@ -198,7 +198,7 @@ class _BaseSmartDevice(UsbHidDriver):
     def _write_fixed_duty(self, cid, duty):
         raise NotImplementedError()
 
-    def set_screen(self, mode, value, **kwargs):
+    def set_screen(self, channel, mode, value, **kwargs):
         """Not supported by this device."""
         raise NotSupportedByDevice()
